requirements.txt
pytest-cov
jupyter
pytest
pandas
pylint
coverage
black
flask
google-cloud
google-cloud-bigquery
google-cloud-storage
google-cloud-core
nltk
